requirements.txt
streamlit
PyMuPDF
faiss-cpu
sentence-transformers
google-generativeai
speechrecognition
pyaudio